上班时间持久化存储

+++ moyubot.py
@@ -93,12 +93,12 @@ class MoyuBot:
         self.__send_html(chat_id, text)
 
     def checkin(self, update, context):
-        chat_id = update.message.chat_id
+        chat_id = int(update.message.chat_id)
         delay: int = 0
         try:
             # 输入小时，在代码级别转化成秒
             def info(cmd): return [i for i in cmd.split(' ') if i != '']
-            delay = 3600*float(info(update.message.text)[1])
+            delay = int(3600*float(info(update.message.text)[1]))
             self.dic[chat_id] = delay
             NoSQLDB().dump()
         except IndexError:
